- Extract GeoJSONGateway to make it easier to find.
